Fix exception in legacy USB support
This was implemented previously, but appears to have been left out in Danny Baumann's tidy up of UsbService
as part of http://review.cyanogenmod.com/6441

When using lagacy USB some of the files in USB_LEGACY_FUNCTIONS_PATH may contain non-numeric data which
causes Integer.valueOf() to throw an exception. If not caught within the loop not all function files
are read and some functions are not set correctly.

Change-Id: Id469ee9e01fca7cc5f33e4976e8e0327a46ca668