Fix for WiMAX signal indicator icon after 2.3.5 merge
The supersonic does not appear to use WIMAX_STATE_CHANGED_ACTION
intent, so the icon does not update from the disconnected state
(iconId = sWimaxDisconnectedImg and mWimaxState = 0 so the
switch(mWimaxState) statement does not change the iconId).
This fix uses the NETWORK_STATE_CHANGED_ACTION intent to set
mWimaxState and mWimaxExtraState so the switch(mWimaxState)
statement sets the iconId for the connection state.

Change-Id: Id2356e10c67d85248d13a82934dbb21a3fe37532