Implement streamlined menu button behavior.
Pressing the home button shows the recent applications list immediately.
This allows switching between recent apps instantly, rather than waiting
for a long press to occur, giving a much more streamlined way to jump
between apps.

The behavior when pressing and releasing and when long-pressing is
"user-habits" compatible with the standard behavior.  Pressing and
releasing quickly still goes to the home menu; long-pressing still pins
the recent menu open.

I think this could be the default, but this is a major UI, so I've
left the standard behavior as the default pending feedback.

Change-Id: Id5253c258fe2fed26fb23f5288b27a4d5f5c27cd
Signed-off-by: Glenn Maynard <glenn@zewt.org>

Conflicts:

	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java