HTC Ril: initial support
Recent htc rom/phone have a newer ril interface which is not compatible
with our rild source and we are force to use the prebuilt.
The rild is not fully compatible with our framework and return an error
during network scan:
RIL_REQUEST_QUERY_AVAILABLE_NETWORKS: invalid response. Got 35 strings,
expected multible of 4

This patch add support for it, just ignoring the last one which is not
used in our framework.

Change-Id: I91c5798402aa4024a29d6328d1942b62dd16693a