Enhance profile framework to support runtime locale changes (1/2)
Before, the localized names of the default profiles were persisted into
the XML file. That way, the locale selected on first boot was used for
determining the profile names.

Now, instead of the name, the resource name is persisted for the default
profiles. That makes the default profile names follow the user's locale,
while custom profiles still use the user selected name.

Change-Id: I11768c3ffea2694aad83bf46709e3d4f5ade521c

Conflicts:

	core/java/android/app/NotificationGroup.java
	core/java/android/app/Profile.java
	core/java/android/app/ProfileManager.java