Fix setting connection status when switching profiles via 3rd party apps.
For switching Wifi and BT, the respective permissions are needed.
Currently, while switching profiles, the caller's identity isn't
cleared, thus the switching action runs under the permission level of
the caller, which is wrong, because the caller shouldn't need to care
about the side effects of profile switching.
To fix this, clear calling identity before profile switching. In order
to now allow clients to pass in profile subclasses with malicious code
in it (and as we don't properly support subclassing the profile classes
anyway), make the profile classes final.