RingLockScreen : Cleanup/Fixes
Addressed the issue where with security on :

Screen on -> Unlock screen -> At security screen turn screen off/on

This was caused by : http://review.cyanogenmod.com/#change,8777

In addition to reverting that i did a few more things to serve the original purpose,
of that patch which was to avoid the lockscreen flash prior to unlocking.
On some phones this was very noticable.

Change-Id: I92569aa2b0a4d2a791f99f5c9f15713cc2c2cbe9

Conflicts:

	core/java/com/android/internal/widget/RingSelector.java
	policy/src/com/android/internal/policy/impl/LockScreen.java