framework: Add a board option to specify a touchpad input id
zeus and zeusc (Xperia play gsm/cdma) have a synaptics touchpad
that is detected as a touchscreen. The InputReader has the facilities
to handle touchpads but the are not used.

This patch adds TARGET_TOUCHPAD_INPUT_DEVICE_ID, a boardconfig option
that allows devices to specify the input device id of the touchpad
and they should work correctly.

This also removes the previous touchpad code that was exclusive to the
zeus board and which will now be moved to the device repository.

Change-Id: Ib1ff12af87e5da8ba334f8fa14fa8ba8ee034800